requirements.txt
Bio==1.8.0
biopython==1.85
datasets==3.2.0
grad_cam==1.5.4
imageio==2.37.0
matplotlib==3.10.3
natsort==8.4.0
numpy==2.2.6
pandas==2.3.0
Pillow==11.2.1
pybedtools==0.11.0
pyBigWig==0.3.24
pyBigWig==0.3.24
pydub==0.25.1
scikit_learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
seqlogo==5.29.9
ssqueezepy==0.6.5
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
wget==3.2
